Tools/Chroma/database-viewer/requirements.txt
flask
chromadb
openai
tiktoken
tree-sitter
tree-sitter-python
python-dotenv
pathspec
